+++ Java/src/main/java/org/example/Main.java
@@ -1,7 +1,5 @@
 package org.example;
-import org.example.Ncc.B0222.RunB0222;
-
-import java.util.Scanner;
+import org.example.Ncc.Feb.B22.RunB0222;
 
 //TIP To <b>Run</b> code, press <shortcut actionId="Run"/> or
 // click the <icon src="AllIcons.Actions.Execute"/> icon in the gutter.
