--- Java/src/main/java/org/example/Ncc/B0222/MultiArrayExample.java
@@ -1,12 +1,0 @@
-package org.example.Ncc.B0222;
-
-public class MultiArrayExample {
-    public static void main(String[] args) {
-        MultiArrayExample mae = new MultiArrayExample();
-
-    }
-
-    public void ex1(){
-
-    }
-}
