+++ Java/src/main/java/org/example/Ncc/Feb/B22/MultiArrayExample.java
@@ -1,0 +1,12 @@
+package org.example.Ncc.Feb.B22;
+
+public class MultiArrayExample {
+    public static void main(String[] args) {
+        MultiArrayExample mae = new MultiArrayExample();
+
+    }
+
+    public void ex1(){
+
+    }
+}
