+++ Java/src/main/java/org/example/Ncc/Feb/B22/RunB0222.java
@@ -1,0 +1,19 @@
+package org.example.Ncc.Feb.B22;
+
+public class RunB0222 {
+    public static void runEx(){
+        ForExample ff = new ForExample();
+        //ff.mulTableEx();
+        //ff.starPrint1();
+        //ff.starPrint2();
+        //ff.mulForEx();
+
+        Test tt = new Test();
+        //tt.test12();
+        tt.problem12();
+
+        DoWhileEx dwe = new DoWhileEx();
+        //dwe.ex1();
+        //dwe.ex2();
+    }
+}
