+++ Java/src/main/java/org/example/Study/Feb/S22/APlusB.java
@@ -1,0 +1,18 @@
+package org.example.Study.Feb.S22;
+
+import java.util.Scanner;
+
+public class APlusB {
+    public static void main(String[] args) {
+        Scanner sc = new Scanner(System.in);
+        int a = sc.nextInt();
+        int b = sc.nextInt();
+
+        System.out.println(a+b);
+        System.out.println(a-b);
+        System.out.println(a*b);
+        System.out.println(a/b);
+
+
+    }
+}
