+++ Java/src/main/java/org/example/Study/Feb/S22/ExchangeTaiYear.java
@@ -1,0 +1,12 @@
+package org.example.Study.Feb.S22;
+
+import java.util.Scanner;
+
+public class ExchangeTaiYear {
+    public static void main(String[] args) {
+        Scanner sc = new Scanner(System.in);
+        int year = sc.nextInt();
+
+        System.out.println(year-543);
+    }
+}
