+++ Java/src/main/java/org/example/Study/Feb/S22/HelloWorld.java
@@ -1,0 +1,7 @@
+package org.example.Study.Feb.S22;
+
+public class HelloWorld {
+    public static void main(String[] args) {
+        System.out.println("Hello World!");
+    }
+}
