+++ Java/src/main/java/org/example/Study/Feb/S22/PrintDog.java
@@ -1,4 +1,4 @@
-package org.example.Study.S0222;
+package org.example.Study.Feb.S22;
 
 public class PrintDog {
     public static void main(String[] args) {
