+++ Java/src/main/java/org/example/Study/Feb/S22/PrintNolam.java
@@ -1,0 +1,12 @@
+package org.example.Study.Feb.S22;
+
+import java.util.Scanner;
+
+public class PrintNolam {
+    public static void main(String[] args) {
+        Scanner sc = new Scanner(System.in);
+        String name= sc.next();
+
+        System.out.println(name +"??!");
+    }
+}
