+++ Java/src/main/java/org/example/Study/Feb/S22/PrintOper.java
@@ -1,0 +1,17 @@
+package org.example.Study.Feb.S22;
+
+import java.util.Scanner;
+
+public class PrintOper {
+    public static void main(String[] args) {
+        Scanner sc = new Scanner(System.in);
+        int a = sc.nextInt();
+        int b = sc.nextInt();
+
+        System.out.println(a+b);
+        System.out.println(a-b);
+        System.out.println(a*b);
+        System.out.println(a/b);
+        System.out.println(a%b);
+    }
+}
