+++ Java/src/main/java/org/example/Study/Feb/S22/PrintRemain.java
@@ -1,0 +1,17 @@
+package org.example.Study.Feb.S22;
+
+import java.util.Scanner;
+
+public class PrintRemain {
+    public static void main(String[] args) {
+        Scanner scanner = new Scanner(System.in);
+        int A = scanner.nextInt();
+        int B = scanner.nextInt();
+        int C = scanner.nextInt();
+
+        System.out.println((A+B)%C);
+        System.out.println(((A%C) + (B%C))%C);
+        System.out.println((A*B)%C);
+        System.out.println(((A%C) * (B%C))%C);
+    }
+}
