+++ Java/src/main/java/org/example/Study/Feb/S22/Test22.java
@@ -1,0 +1,5 @@
+package org.example.Study.Feb.S22;
+
+public class Test22 {
+
+}
