--- Java/src/main/java/org/example/Study/S0222/ExchangeTaiYear.java
@@ -1,12 +1,0 @@
-package org.example.Study.S0222;
-
-import java.util.Scanner;
-
-public class ExchangeTaiYear {
-    public static void main(String[] args) {
-        Scanner sc = new Scanner(System.in);
-        int year = sc.nextInt();
-
-        System.out.println(year-543);
-    }
-}
