--- Java/src/main/java/org/example/Study/S0222/HelloWorld.java
@@ -1,7 +1,0 @@
-package org.example.Study.S0222;
-
-public class HelloWorld {
-    public static void main(String[] args) {
-        System.out.println("Hello World!");
-    }
-}
