--- Java/src/main/java/org/example/Study/S0222/PrintNolam.java
@@ -1,12 +1,0 @@
-package org.example.Study.S0222;
-
-import java.util.Scanner;
-
-public class PrintNolam {
-    public static void main(String[] args) {
-        Scanner sc = new Scanner(System.in);
-        String name= sc.next();
-
-        System.out.println(name +"??!");
-    }
-}
