--- Java/src/main/java/org/example/Study/S0222/PrintOper.java
@@ -1,17 +1,0 @@
-package org.example.Study.S0222;
-
-import java.util.Scanner;
-
-public class PrintOper {
-    public static void main(String[] args) {
-        Scanner sc = new Scanner(System.in);
-        int a = sc.nextInt();
-        int b = sc.nextInt();
-
-        System.out.println(a+b);
-        System.out.println(a-b);
-        System.out.println(a*b);
-        System.out.println(a/b);
-        System.out.println(a%b);
-    }
-}
