--- Java/src/main/java/org/example/Study/S0222/PrintRemain.java
@@ -1,17 +1,0 @@
-package org.example.Study.S0222;
-
-import java.util.Scanner;
-
-public class PrintRemain {
-    public static void main(String[] args) {
-        Scanner scanner = new Scanner(System.in);
-        int A = scanner.nextInt();
-        int B = scanner.nextInt();
-        int C = scanner.nextInt();
-
-        System.out.println((A+B)%C);
-        System.out.println(((A%C) + (B%C))%C);
-        System.out.println((A*B)%C);
-        System.out.println(((A%C) * (B%C))%C);
-    }
-}
