--- Java/src/main/java/org/example/Study/S0222/Test22.java
@@ -1,5 +1,0 @@
-package org.example.Study.S0222;
-
-public class Test22 {
-
-}
